Make ToDto and ToModel virtual
Users of the Mapper are encouraged to use Properties and name them the same across Models and Dtos. If for some reason, this is too restrictive then the methods can be overriden. The base method can be used to map the matching Properties, then everything else can be mapped manually in the overridden method.